lab12_1.cpp

+++ lab12_1.cpp
@@ -1,6 +1,6 @@
 #include<iostream>
 using namespace std;
-int fibonacci(int x);
+long long int fibonacci(int x);
 
 int main(){
     clock_t start = clock();
@@ -11,7 +11,7 @@ int main(){
     return 0;
 }
 
-int fibonacci(int x){
+long long int fibonacci(int x){
     if(x == 0)
         return 0;
     if(x == 1)
